requirements.txt
pytest
pytest-cov
coverage
Pillow